+++ src/oph_workflow/oph_workflow_parsing.c
@@ -83,10 +83,10 @@ int oph_workflow_load(char *json_string, const char *username, const char *ip_ad
 
 	//unpack global vars
 	char *name = NULL, *author = NULL, *abstract = NULL, *sessionid = NULL, *exec_mode = NULL, *ncores = NULL, *cwd = NULL, *cdd = NULL, *cube = NULL, *callback_url = NULL, *on_error =
-	    NULL, *command = NULL, *on_exit = NULL, *run = NULL, *output_format = NULL, *host_partition = NULL, *url = NULL, *nhosts = NULL, *nthreads = NULL, *project = NULL;
-	json_unpack(jansson, "{s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s}", "name", &name, "author", &author, "abstract", &abstract, "sessionid", &sessionid,
+	    NULL, *command = NULL, *on_exit = NULL, *run = NULL, *output_format = NULL, *host_partition = NULL, *url = NULL, *nhosts = NULL, *nthreads = NULL, *project = NULL, *save = NULL;
+	json_unpack(jansson, "{s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s,s?s}", "name", &name, "author", &author, "abstract", &abstract, "sessionid", &sessionid,
 		    "exec_mode", &exec_mode, "ncores", &ncores, "cwd", &cwd, "cdd", &cdd, "cube", &cube, "callback_url", &callback_url, "on_error", &on_error, "command", &command, "on_exit", &on_exit,
-		    "run", &run, "output_format", &output_format, "host_partition", &host_partition, "url", &url, "nhost", &nhosts, "nthreads", &nthreads, "project", &project);
+		    "run", &run, "output_format", &output_format, "host_partition", &host_partition, "url", &url, "nhost", &nhosts, "nthreads", &nthreads, "project", &project, "save", &save);
 
 	//add global vars
 	if (!name || !author || !abstract) {
@@ -248,6 +248,18 @@ int oph_workflow_load(char *json_string, const char *username, const char *ip_ad
 			return OPH_WORKFLOW_EXIT_BAD_PARAM_ERROR;
 		}
 	}
+	(*workflow)->save = 1;	// Default value (yes)
+	if (save && strlen(save)) {
+		if (!strcmp(save, OPH_WORKFLOW_NO))
+			(*workflow)->save = 0;
+		else if (strcmp(save, OPH_WORKFLOW_YES)) {
+			oph_workflow_free(*workflow);
+			if (jansson)
+				json_decref(jansson);
+			pmesg(LOG_ERROR, __FILE__, __LINE__, "error in parsing parameter 'save'\n");
+			return OPH_WORKFLOW_EXIT_BAD_PARAM_ERROR;
+		}
+	}
 	(*workflow)->output_format = 0;
 	if (output_format && strlen(output_format)) {
 		if (!strcmp(output_format, OPH_WORKFLOW_COMPACT))
@@ -312,8 +324,8 @@ int oph_workflow_load(char *json_string, const char *username, const char *ip_ad
 			return OPH_WORKFLOW_EXIT_GENERIC_ERROR;
 		}
 		//unpack name and operator
-		char *name = NULL, *operator= NULL, *on_error_task = NULL, *on_exit_task = NULL, *run_task = NULL;
-		json_unpack(task, "{s?s,s?s,s?s,s?s,s?s}", "name", &name, "operator", &operator, "on_error", &on_error_task, "on_exit", &on_exit_task, "run", &run_task);
+		char *name = NULL, *operator= NULL, *on_error_task = NULL, *on_exit_task = NULL, *run_task = NULL, *save_task = NULL;
+		json_unpack(task, "{s?s,s?s,s?s,s?s,s?s,s?s}", "name", &name, "operator", &operator, "on_error", &on_error_task, "on_exit", &on_exit_task, "run", &run_task, "save", &save_task);
 
 		//add name and operator
 		if (!name || !operator) {
@@ -616,6 +628,21 @@ int oph_workflow_load(char *json_string, const char *username, const char *ip_ad
 				if (jansson)
 					json_decref(jansson);
 				pmesg(LOG_ERROR, __FILE__, __LINE__, "error in parsing parameter 'run'\n");
+				return OPH_WORKFLOW_EXIT_BAD_PARAM_ERROR;
+			}
+		}
+
+		(*workflow)->tasks[i].save = 1;	// Default value (yes)
+		if (!save_task)
+			save_task = save;
+		if (save_task && strlen(save_task)) {
+			if (!strcmp(save_task, OPH_WORKFLOW_NO))
+				(*workflow)->tasks[i].save = 0;
+			else if (strcmp(save_task, OPH_WORKFLOW_YES)) {
+				oph_workflow_free(*workflow);
+				if (jansson)
+					json_decref(jansson);
+				pmesg(LOG_ERROR, __FILE__, __LINE__, "error in parsing parameter 'save'\n");
 				return OPH_WORKFLOW_EXIT_BAD_PARAM_ERROR;
 			}
 		}