+++ HISTORY.md
@@ -3,12 +3,17 @@
 
 ### Fixed:
 
+- Bug in handling the arguments 'enable' and 'disable' of OPH_SERVICE
 - Bug in transfer metadata into accounting table
 - Bug in handling cube pids for final operation
 
 ### Added:
 
 - New argument 'WORKING_DIR' in main configuration file
+
+### Changed:
+
+- Activate service without authorization procedure in case authorization directory is not found
 
 
 ## v1.6.1 - 2021-03-16