+++ HISTORY.md
@@ -1,5 +1,5 @@
 
-## Next release
+## v1.6.2 - 2021-08-02
 
 ### Fixed:
 
@@ -9,10 +9,12 @@
 
 ### Added:
 
+- Option to enable and disable JSON response saving [#56](https://github.com/OphidiaBigData/ophidia-server/pull/56)
 - New argument 'WORKING_DIR' in main configuration file
 
 ### Changed:
 
+- File check of OPH_WAIT from 'stat' to OPH_FS [#57](https://github.com/OphidiaBigData/ophidia-server/pull/57)
 - Activate service without authorization procedure in case authorization directory is not found